+++ experiments/evaluation/policy_eval.py
@@ -3,22 +3,19 @@
 import os
 
 import tensorflow as tf
-from tf_agents.environments import tf_py_environment
-from tf_agents.environments import wrappers
-
-from optfuncs import numpy_functions as npf
-
-from src.environments import py_function_env as py_fun_env
-from src import config
 
 from experiments.evaluation import utils as eval_utils
+from optfuncs import tensorflow_functions as tff
+from sarlopt import config
+from sarlopt.environments import tf_function_env as tf_fun_env
 
 POLICIES_DIR = config.POLICIES_DIR
 
 if __name__ == '__main__':
-  function = npf.Sphere()
+  function = tff.Sphere()
   dims = 30
   steps = 500
+  seed = 10000
 
   policy_dir = os.path.join(POLICIES_DIR, 'Td3Agent')
   policy_dir = os.path.join(policy_dir, f'{dims}D')
@@ -26,10 +23,10 @@ if __name__ == '__main__':
 
   saved_pol = tf.compat.v2.saved_model.load(policy_dir)
 
-  env = py_fun_env.PyFunctionEnv(function, dims)
-  env = wrappers.TimeLimit(env, duration=steps)
-
-  tf_eval_env = tf_py_environment.TFPyEnvironment(environment=env)
+  tf_eval_env = tf_fun_env.TFFunctionEnv(function=function,
+                                         dims=dims,
+                                         duration=steps,
+                                         seed=seed)
 
   # tf.config.run_functions_eagerly(True)
   eval_utils.evaluate_agent(eval_env=tf_eval_env,
