+++ experiments/training/reinforce_baseline.py
@@ -14,9 +14,9 @@ from tf_agents.metrics import tf_metrics
 
 from optfuncs import tensorflow_functions as tff
 
-from src.environments import tf_function_env as tf_fun_env
-from src.typing.types import LayerParam
-from src.metrics import tf_custom_metrics
+from sarlopt.environments import tf_function_env as tf_fun_env
+from sarlopt.typing.types import LayerParam
+from sarlopt.metrics import tf_custom_metrics
 
 from experiments.evaluation import utils as eval_utils
 from experiments.training import utils as training_utils
