+++ sarlopt/environments/tf_function_env_v2_test.py
@@ -8,8 +8,8 @@ from tf_agents.typing import types
 from optfuncs import core
 from optfuncs import tensorflow_functions as tff
 
-from src.environments.tf_function_env_v2 import TFFunctionEnvV2
-from src.utils.random import functions as rand_fn
+from sarlopt.environments.tf_function_env_v2 import TFFunctionEnvV2
+from sarlopt.utils.random import functions as rand_fn
 
 
 class DummyFunction(tff.TensorflowFunction):
