+++ sarlopt/networks/linear_actor_network.py
